Set correct licence
I'm uncertain whether this is legally correct, but it seems to me
the intention was to set it to AGPLv3 from the get-go.

Not to mention, OctoPrint itself is AGPL and my understanding is
that all dependencies and plugins it uses also need to be AGPL.

I could be wrong, but the codebase might be completely
unregonisable when I finish refactoring. If I'm wrong, I'll revert this commit.